requirements.txt
transformers 
datasets 
accelerate 
pyyaml
torch
numpy